More bugs is fixed
