+++ client/public/index.html
@@ -19,14 +19,14 @@
     <script src="%PUBLIC_URL%/js/jquery.min.js"></script>
     <script src="%PUBLIC_URL%/css/bootstrap4/popper.js"></script>
     <script src="%PUBLIC_URL%/css/bootstrap4/bootstrap.min.js"></script>
+    <script src="%PUBLIC_URL%/js/custom.js"></script>       
   </head>
   <body>
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
     <div id="super_container"></div>
-    
-    
+
     <script src="%PUBLIC_URL%/plugins/greensock/TweenMax.min.js"></script>
     <script src="%PUBLIC_URL%/plugins/greensock/TimelineMax.min.js"></script>
     <script src="%PUBLIC_URL%/plugins/scrollmagic/ScrollMagic.min.js"></script>
@@ -34,8 +34,6 @@
     <script src="%PUBLIC_URL%/plugins/greensock/ScrollToPlugin.min.js"></script>
     <script src="%PUBLIC_URL%/plugins/OwlCarousel2-2.2.1/owl.carousel.js"></script>
     <script src="%PUBLIC_URL%/plugins/slick-1.8.0/slick.js"></script>
-    <script src="%PUBLIC_URL%/plugins/easing/easing.js"></script>
-    <script src="%PUBLIC_URL%/js/custom.js"></script>
-    <script src="%PUBLIC_URL%/js/custom_main.js"></script>
+    <script src="%PUBLIC_URL%/plugins/easing/easing.js"></script>                         
   </body>
 </html>
