+++ client/src/components/Main/MainComStyle.css
@@ -1,0 +1,3 @@
+.modal-open {
+    overflow: auto;
+}
