+++ client/src/utils/storage.js
@@ -16,12 +16,24 @@ export function getFromStorage(key) {
 
 export function setInStorage(key, obj) {
     if (!key) {
-        console.error('Lỗi: không có tìm thấy key');
+        console.error('Lỗi: không có key cần tìm');
     }
 
     try {
         localStorage.setItem(key, JSON.stringify(obj));
     } catch (err) {
         console.error(err);
+    }
+}
+
+export function removeFromStorage(key) {
+    if (!key) {
+        console.log('Lỗi: không có key cần tìm');
+    }
+
+    try {
+        localStorage.removeItem(key);
+    } catch (err) {
+        console.log(err);
     }
 }
